fix: Windows linker error and improve RemeshResult handling
- Replace MMGS_Get_triangleQuality with compute_triangle_quality
  (MMGS function not exported from DLL on Windows)
- Remove static_cast<int> from result dict values (let pybind11
  handle MMG5_int to Python conversion, fixes potential overflow)
- Fix quality_improvement property to return 1.0 when both before
  and after are zero (no change case)
- Wrap remesh_lagrangian and remesh_levelset to return RemeshResult
- Add tests for remesh_levelset across all mesh types
- Add tests for quality_improvement edge cases

+++ src/bindings/mmg_mesh.cpp
@@ -1259,14 +1259,14 @@ py::dict MmgMesh::remesh(const py::dict &options) {
       tets_after > 0 ? quality_sum_after / tets_after : 0.0;
 
   py::dict result;
-  result["vertices_before"] = static_cast<int>(verts_before);
-  result["vertices_after"] = static_cast<int>(verts_after);
-  result["elements_before"] = static_cast<int>(tets_before);
-  result["elements_after"] = static_cast<int>(tets_after);
-  result["triangles_before"] = static_cast<int>(tris_before);
-  result["triangles_after"] = static_cast<int>(tris_after);
-  result["edges_before"] = static_cast<int>(edges_before);
-  result["edges_after"] = static_cast<int>(edges_after);
+  result["vertices_before"] = verts_before;
+  result["vertices_after"] = verts_after;
+  result["elements_before"] = tets_before;
+  result["elements_after"] = tets_after;
+  result["triangles_before"] = tris_before;
+  result["triangles_after"] = tris_after;
+  result["edges_before"] = edges_before;
+  result["edges_after"] = edges_after;
   result["quality_min_before"] = quality_min_before;
   result["quality_min_after"] = quality_min_after;
   result["quality_mean_before"] = quality_mean_before;
@@ -1336,14 +1336,14 @@ py::dict MmgMesh::remesh_lagrangian(const py::array_t<double> &displacement,
       tets_after > 0 ? quality_sum_after / tets_after : 0.0;
 
   py::dict result;
-  result["vertices_before"] = static_cast<int>(verts_before);
-  result["vertices_after"] = static_cast<int>(verts_after);
-  result["elements_before"] = static_cast<int>(tets_before);
-  result["elements_after"] = static_cast<int>(tets_after);
-  result["triangles_before"] = static_cast<int>(tris_before);
-  result["triangles_after"] = static_cast<int>(tris_after);
-  result["edges_before"] = static_cast<int>(edges_before);
-  result["edges_after"] = static_cast<int>(edges_after);
+  result["vertices_before"] = verts_before;
+  result["vertices_after"] = verts_after;
+  result["elements_before"] = tets_before;
+  result["elements_after"] = tets_after;
+  result["triangles_before"] = tris_before;
+  result["triangles_after"] = tris_after;
+  result["edges_before"] = edges_before;
+  result["edges_after"] = edges_after;
   result["quality_min_before"] = quality_min_before;
   result["quality_min_after"] = quality_min_after;
   result["quality_mean_before"] = quality_mean_before;
@@ -1412,14 +1412,14 @@ py::dict MmgMesh::remesh_levelset(const py::array_t<double> &levelset,
       tets_after > 0 ? quality_sum_after / tets_after : 0.0;
 
   py::dict result;
-  result["vertices_before"] = static_cast<int>(verts_before);
-  result["vertices_after"] = static_cast<int>(verts_after);
-  result["elements_before"] = static_cast<int>(tets_before);
-  result["elements_after"] = static_cast<int>(tets_after);
-  result["triangles_before"] = static_cast<int>(tris_before);
-  result["triangles_after"] = static_cast<int>(tris_after);
-  result["edges_before"] = static_cast<int>(edges_before);
-  result["edges_after"] = static_cast<int>(edges_after);
+  result["vertices_before"] = verts_before;
+  result["vertices_after"] = verts_after;
+  result["elements_before"] = tets_before;
+  result["elements_after"] = tets_after;
+  result["triangles_before"] = tris_before;
+  result["triangles_after"] = tris_after;
+  result["edges_before"] = edges_before;
+  result["edges_after"] = edges_after;
   result["quality_min_before"] = quality_min_before;
   result["quality_min_after"] = quality_min_after;
   result["quality_mean_before"] = quality_mean_before;
